pyproject.toml
[project]
name = "synthetic-pdf-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=24.10.0",
    "fpdf2>=2.8.1",
    "ftfy>=6.3.1",
    "nanoid>=2.0.0",
    "openai>=1.54.3",
    "python-dotenv>=1.0.1",
]
